Resolve problem with libphidget22 and Debian package naming rules
The Debian package naming convention will rename a package after the
shared library name if it is the only thing present in the package.

This resulted in libphidget22-upstream being renamed to libphidget22
and there being 2 packages present with that name.

As a workaround we set DEBIAN_NAMES to zero to disable the Debian
naming rules for that package.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>