openvslam:  Disable compiler option overrides
The CMakeLists.txt for FBoW tries to automatically set the compiler
options.

This interferes with the compiler options set by the build system.  It
causes FBoW to fail to find iostream because --sysroot is not set.

We patch out this part of FBoW CMakeLists.txt to prevent this from
happening.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>